Made it work.
User can hit the activate endpoint and it will work!
When activate endpoint is called:
It checks the user and finds the IMAP config for that user. It takes the IMAP config and makes a connection to the email provider. If connection is established, it gets all the unread emails and sends them to BullMQ. Email analysis takes new emails from queue and processes it and saves in another queue. The tool service takes new processed emails from the queue and runs any tool the email type fits.